requirements.txt
rich
logging
click
pandas
firecrawl-py
python-dotenv
git+https://gitlab.com/xgrg/bibtex2docs.git
google-generativeai
pytest
pytest-mock
pydantic
